+++ logic/backtest/__init__.py
@@ -1,1 +1,17 @@
-"""backtest package: contains runner for backtesting."""
+"""Backtest package: runners and reporting utilities."""
+
+from .country_runner import CountryBacktestResult, DEFAULT_TEST_MONTHS_RANGE, run_country_backtest
+from .reporting import (
+    dump_backtest_log,
+    format_period_return_with_listing_date,
+    print_backtest_summary,
+)
+
+__all__ = [
+    "CountryBacktestResult",
+    "DEFAULT_TEST_MONTHS_RANGE",
+    "run_country_backtest",
+    "dump_backtest_log",
+    "format_period_return_with_listing_date",
+    "print_backtest_summary",
+]
