+++ logic/tune/__init__.py
@@ -1,1 +1,5 @@
-"""tune package: contains parameter tuning runner."""
+"""Utilities for running parameter tuning workflows."""
+
+from .runner import run_country_tuning
+
+__all__ = ["run_country_tuning"]
