+++ requirements.txt
@@ -38,7 +38,7 @@ charset-normalizer==3.4.3
     # via requests
 click==8.2.1
     # via streamlit
-contourpy==1.3.3
+contourpy==1.3.2
     # via matplotlib
 cron-descriptor==2.0.6
     # via -r requirements.in
